+++ eras/allegra/impl/test/Test/Cardano/Ledger/Allegra/BinarySpec.hs
@@ -11,6 +11,7 @@ import Test.Cardano.Ledger.Allegra.Arbitrary ()
 import Test.Cardano.Ledger.Allegra.TreeDiff ()
 import Test.Cardano.Ledger.Common
 import Test.Cardano.Ledger.Core.Binary (specUpgrade)
+import Test.Cardano.Ledger.Core.Binary as Binary (decoderEquivalenceCoreEraTypesSpec)
 import Test.Cardano.Ledger.Core.Binary.RoundTrip (RuleListEra (..))
 import Test.Cardano.Ledger.Shelley.Binary.RoundTrip (roundTripShelleyCommonSpec)
 
@@ -19,6 +20,8 @@ spec = do
   specUpgrade @AllegraEra def
   describe "RoundTrip" $ do
     roundTripShelleyCommonSpec @AllegraEra
+  describe "DecCBOR instances equivalence" $ do
+    Binary.decoderEquivalenceCoreEraTypesSpec @AllegraEra
 
 instance RuleListEra AllegraEra where
   type