+++ eras/conway/impl/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.19.0.0
 
+* Add `DecCBOR` instance for `ConwayTxBody`
 * Remove `ConwayMempoolPredFailure` and `ConwayMempoolEvent`
 * Switch to `MEMPOOL` rule to be the entry point for `ApplyTx` instead of `LEDGER` and invert their
   ivocation.