+++ eras/shelley/impl/CHANGELOG.md
@@ -2,6 +2,15 @@
 
 ## 1.16.0.0
 
+* Add `DecCBOR` instances for:
+  * `ShelleyTxWits`
+  * `ShelleyTxAuxData`
+  * `ShelleyTxBody`
+  * `ShelleyTx`
+  * `ShelleyTxSeq`
+* Add `indexLookupSeq`
+* Add `segWitTx`
+* Rename `segwitTx` to `segWitAnnTx`
 * Restrict the monad of `applyTx` and `reapllyTx` to `Either` from abstract `MonadError`
 * Remove `applyTxOpts` in favor of new `applyTxValidation` function in `ApplyTx`
 * Move `reapplyTx` outside of the `ApplyTx` type class.