+++ libs/cardano-ledger-binary/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.6.0.0
 
+* Add `ToCBOR` instance for `PV1.Data`
 * Add `DecCBOR` instance for `Annotated a ByteString`
 * Add `originalBytesExpectedFailureMessage` needed for testing
 * Add `decodeListLikeWithCountT`
@@ -31,6 +32,12 @@
 
 ### `testlib`
 
+* Add:
+  * `decoderEquivalenceSpec`
+  * `decoderEquivalenceExpectation`
+  * `decoderEquivalenceProp`
+  * `cddlDecoderEquivalenceSpec`
+  * `huddleDecoderEquivalenceSpec`
 * Re-export types `Doc` and `AnsiStyle` in `Test.Cardano.Ledger.Binary.TreeDiff`
 * `diffExpr` and `diffExprCompact` changed type signature
 * Add `diffExprString` and `diffExprCompactString`, which replace the old implementations